replaced automapper with mapster

+++ src/Domain/Fw.Domain.Ams/Contracts/SubmitInvoiceLine.cs
@@ -7,6 +7,6 @@ public record SubmitInvoiceLine : Request<InvoiceLineSubmitted>
 {
     public Guid InvoiceId { get; set; }
     public int LineNumber { get; set; }
-    public int LineQuantity { get; set; }
+    public decimal LineQuantity { get; set; }
     public string LineDescription { get; set; }
 }
